+++ scripts/install/php-fpm.sh
@@ -17,15 +17,10 @@ sed -i "s|;date.timezone =.*|date.timezone = ${PHP_TIMEZONE}|g" /etc/php/7.0/fpm
 sed -i "s|;date.timezone =.*|date.timezone = ${PHP_TIMEZONE}|g" /etc/php/7.0/cli/php.ini
 
 mkdir -p /run/php
+chown -R www-data:www-data /run/php
 
 # Logs
-mkdir -p $DATA_DIRECTORY/logs/php-fpm
-mkdir -p $DATA_DIRECTORY/logs/php-cli
-sed -i "s|;error_log = php_errors.log|error_log = ${DATA_DIRECTORY}/logs/php-fpm/errors.log|g" /etc/php5/fpm/php.ini
-sed -i "s|;error_log = php_errors.log|error_log = ${DATA_DIRECTORY}/logs/php-cli/errors.log|g" /etc/php5/cli/php.ini
-
-# Loading permissions
-chown -R www-data:www-data \
-    /run/php \
-    ${DATA_DIRECTORY}/logs/php-fpm \
-    ${DATA_DIRECTORY}/logs/php-cli
+mkdir -p ${DATA_DIRECTORY}/logs/php-fpm
+mkdir -p ${DATA_DIRECTORY}/logs/php-cli
+sed -i "s|;error_log = php_errors.log|error_log = ${DATA_DIRECTORY}/logs/php-fpm-errors.log|g" /etc/php/7.0/fpm/php.ini
+sed -i "s|;error_log = php_errors.log|error_log = ${DATA_DIRECTORY}/logs/php-cli-errors.log|g" /etc/php/7.0/cli/php.ini